requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==35.0.0
Django==3.2.9
django-celery-beat==2.2.1
django-rest-knox==4.1.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
idna==3.3
json-excel-converter==1.3.0
kombu==5.2.1
prompt-toolkit==3.0.22
pycparser==2.21
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
XlsxWriter==3.0.2
